+++ lib/s_future_button.dart
@@ -3,4 +3,4 @@
 /// This library exports all public APIs for the future_button package.
 library;
 
-export 'src/future_button.dart';
+export 'src/s_future_button.dart';